Extract `Account` model to the dedicated file
